+++ WalletWasabi.Fluent/Behaviors/Suggestion.cs
@@ -1,0 +1,17 @@
+using System.Reactive;
+using ReactiveUI;
+
+namespace WalletWasabi.Fluent.Behaviors;
+
+public class Suggestion : ReactiveObject
+{
+	public Suggestion(string text, Action onAccept)
+	{
+		Text = text;
+		AcceptHintCommand = ReactiveCommand.Create(onAccept);
+	}
+
+	public string Text { get; }
+
+	public ReactiveCommand<Unit, Unit> AcceptHintCommand { get; set; }
+}
